♻️ Rename db folder to database
