+++ drizzle.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "drizzle-kit";
 
 export default defineConfig({
-  schema: "./db/schema.ts",
+  schema: "./database/schema.ts",
   out: "./drizzle",
   dialect: "sqlite",
   driver: "expo",
